// src/app/models/enums/blood.type.ts
/**
 * O enum que representa os tipos sanguineos
 */
export enum BloodType {

  /**
   * O tipo sanguineo A+
   */
  A_PLUS = 'A+',

  /**
   * O tipo sanguineo A-
   */
  A_MINUS = 'A-',

  /**
   * O tipo sanguineo B+
   */
  B_PLUS = 'B+',

  /**
   * O tipo sanguineo B-
   */
  B_MINUS = 'B-',

  /**
   * O tipo sanguineo O+
   */
  O_PLUS = 'O+',

  /**
   * O tipo sanguineo O-
   */
  O_MINUS = 'O-',

  /**
   * O tipo sanguineo AB+
   */
  AB_PLUS = 'AB+',

  /**
   * O tipo sanguineo AB-
   */
  AB_MINUS = 'AB-',

}